Split SYCL up
